+++ db/a_co_db_table_base.class.php
@@ -252,6 +252,9 @@ abstract class A_CO_DB_Table_Base {
         } else {
             if (isset($ids) && is_array($ids) && count($ids)) {
                 $ret = in_array($my_read_item, $ids);
+                if (!$ret && $this->get_access_object()->get_login_id()) {
+                    $ret = (1 == $my_read_item);  // Logged-in users can read 1s.
+                }
             }
         }